+++ src/app/exhibition-info/osaka/all/page.tsx
@@ -1,7 +1,7 @@
 import type { Metadata } from 'next';
 import Image from 'next/image';
 
-export const metadata: Metadata = { title: '오사카 전체 맵 - 기획전 정보' };
+export const metadata: Metadata = { title: '오사카 전체 지도 - 기획전 정보' };
 
 
 
@@ -10,17 +10,17 @@ export default function ExhibitionInfoPage() {
     <main>
       <header className='header'>
         <div className="container" style={{ textAlign: 'center' }}>
-          <h1 className="header-title">오사카 전체 지도</h1>
-          <p className="header-subtitle">2025년 8월 9일 (토) ~ 8월 11일 (월)</p>
+          <h1 className="header-title">전체 지도</h1>
+          <p className="header-subtitle">오사카 기획전 - 매지컬 미라이 2025</p>
         </div>
       </header>
 
       {/* <div className="info-section"> */}
         {/* <h1 className="info-section-title">전체 지도</h1> */}
-        <div className="image-map-container">
+        <div className="image-map-container" style={{ aspectRatio: 1500 / 639 }}>
           <Image
             src="https://github.com/kuroka3/MM-Info/blob/master/.github/osaka_assets/map_osaka_all_kr.png?raw=true"
-            alt="오사카 전체 맵"
+            alt="오사카 전체 지도"
             fill
             priority
             style={{ objectFit: 'contain', borderRadius: '16px' }}
@@ -34,7 +34,7 @@ export default function ExhibitionInfoPage() {
               width: '14.33%',
               height: '34.90%',
             }}
-          ></a>
+          />
           <a
             href="/exhibition-info/osaka/4"
             className="image-map-link"
@@ -44,7 +44,7 @@ export default function ExhibitionInfoPage() {
               width: '21.53%',
               height: '34.90%',
             }}
-          ></a>
+          />
           <a
             href="/exhibition-info/osaka/3"
             className="image-map-link"
@@ -54,7 +54,7 @@ export default function ExhibitionInfoPage() {
               width: '16.73%',
               height: '39.28%',
             }}
-          ></a>
+          />
         </div>
       {/* </div> */}
     </main>
