+++ src/app/globals.css
@@ -1552,19 +1552,10 @@ body {
   margin-bottom: 4rem;
 }
 
-.info-section-title {
-  font-size: 1.5rem;
-  font-weight: 700;
-  margin-bottom: 16px;
-  color: var(--text-primary);
-  text-align: center;
-}
-
 .image-map-container {
   margin-top: 120px;
   position: relative;
   width: 100%;
-  aspect-ratio: 1500 / 639;
 }
 
 .image-map-link {
@@ -1576,3 +1567,75 @@ body {
   background-color: rgba(255, 255, 255, 0.4);
 }
 
+.info-nav ul {
+  display: grid;
+  grid-template-columns: repeat(3, 1fr);
+  gap: 12px;
+  width: 100%;
+  margin: 2rem 0;
+  list-style: none;
+  padding: 0;
+}
+.info-nav ul:has(li:only-child) {
+  grid-template-columns: 1fr;
+  width: calc(100% / 3);
+  margin-left: auto;
+  margin-right: auto;
+}
+.info-nav ul:has(li:nth-child(2):last-child) {
+  grid-template-columns: repeat(2, 1fr);
+  width: calc(100% * 2 / 3);
+  margin-left: auto;
+  margin-right: auto;
+}
+.info-nav a {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100%;
+  padding: 12px 16px;
+  text-align: center;
+  border-radius: 12px;
+  background: var(--glass-bg);
+  border: 1px solid #fff;
+  backdrop-filter: blur(10px);
+  text-decoration: none;
+  color: var(--text-primary);
+  font-size: clamp(1rem, 2.5vw, 1.2rem);
+  transition: background 0.2s, border-color 0.2s;
+}
+.info-nav a:hover {
+  background: var(--surface);
+  border-color: #fff;
+}
+
+@media (max-width: 600px) {
+  .info-nav ul {
+    grid-template-columns: 1fr;
+  }
+}
+
+.booth-section {
+  background: var(--surface);
+  border: 1px solid var(--border);
+  border-radius: 24px;
+  padding: 32px;
+  margin-top: 2rem;
+  margin-bottom: 2rem;
+}
+
+.booth-section-title {
+  font-size: 2rem;
+  font-weight: 700;
+  margin-bottom: 16px;
+  color: var(--text-primary);
+  text-align: center;
+}
+
+.booth-image {
+  max-width: 100%;
+  border-radius: 12px;
+  margin: 8px 0;
+  object-fit: contain;
+  border-radius: 8px;
+}
